Moving FrozenDefaultDict to utils
