+++ src/tqec/circuit/generation_test.py
@@ -2,10 +2,10 @@ import stim
 
 from tqec.circuit.generation import generate_circuit
 from tqec.plaquette.enums import PlaquetteOrientation
-from tqec.plaquette.frozendefaultdict import FrozenDefaultDict
 from tqec.plaquette.library import make_css_surface_code_plaquette
 from tqec.plaquette.plaquette import Plaquettes
 from tqec.templates.indices._testing import FixedTemplate
+from tqec.utils.frozendefaultdict import FrozenDefaultDict
 
 
 def test_generate_circuit_one_plaquette() -> None:
