+++ src/tqec/compile/detectors/compute_test.py
@@ -22,7 +22,6 @@ from tqec.compile.detectors.database import DetectorDatabase
 from tqec.compile.detectors.detector import Detector
 from tqec.enums import Basis
 from tqec.exceptions import TQECException
-from tqec.plaquette.frozendefaultdict import FrozenDefaultDict
 from tqec.plaquette.library.css import make_css_surface_code_plaquette
 from tqec.plaquette.library.empty import empty_square_plaquette
 from tqec.plaquette.plaquette import Plaquettes
@@ -31,6 +30,7 @@ from tqec.templates.indices._testing import FixedTemplate
 from tqec.templates.indices.layout import LayoutTemplate
 from tqec.templates.indices.qubit import QubitTemplate
 from tqec.templates.indices.subtemplates import SubTemplateType
+from tqec.utils.frozendefaultdict import FrozenDefaultDict
 
 
 @pytest.fixture(name="alternating_subtemplate")
