+++ src/tqec/plaquette/plaquette.py
@@ -10,10 +10,10 @@ from typing_extensions import override
 from tqec.circuit.schedule import ScheduledCircuit
 from tqec.exceptions import TQECException
 from tqec.plaquette.enums import PlaquetteOrientation
-from tqec.plaquette.frozendefaultdict import FrozenDefaultDict
 from tqec.plaquette.qubit import PlaquetteQubits
 from tqec.position import PhysicalQubitPosition2D
 from tqec.scale import LinearFunction, round_or_fail
+from tqec.utils.frozendefaultdict import FrozenDefaultDict
 
 
 @dataclass(frozen=True)
