+++ src/tqec/templates/library/memory.py
@@ -1,5 +1,4 @@
 from tqec.enums import Basis
-from tqec.plaquette.frozendefaultdict import FrozenDefaultDict
 from tqec.plaquette.rpng import RPNGDescription
 from tqec.templates.enums import ZObservableOrientation
 from tqec.templates.indices.qubit import (
@@ -7,6 +6,7 @@ from tqec.templates.indices.qubit import (
     QubitTemplate,
     QubitVerticalBorders,
 )
+from tqec.utils.frozendefaultdict import FrozenDefaultDict
 
 
 def get_memory_qubit_raw_template() -> QubitTemplate:
