+++ src/tqec/templates/rpng.py
@@ -3,11 +3,11 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Generic, TypeVar
 
-from tqec.plaquette.frozendefaultdict import FrozenDefaultDict
 from tqec.plaquette.rpng import RPNGDescription
 from tqec.position import PlaquettePosition2D, Shape2D, Shift2D
 from tqec.scale import Scalable2D
 from tqec.templates.indices.base import Template
+from tqec.utils.frozendefaultdict import FrozenDefaultDict
 
 T = TypeVar("T", bound=Template, covariant=True)
